stabilityai-stable-diffusion-xl/requirements.txt
diffusers
transformers
accelerate
safetensors
xformers==0.0.23
runpod==1.7.9
numpy==1.26.3
scipy
triton
huggingface-hub
hf_transfer
hf_xet
setuptools
boto3
pillow
datetime
pyOpenSSL
cryptography
nanoid